+++ src/domain/entities/Bus.cpp
@@ -14,7 +14,7 @@ namespace tdc::domain::entities {
         serviceType = Express;
         driver = nullptr;
     };
-    Bus::Bus(string _id, int8_t _seats, vector<BusAmenities> _amenities, ServiceType _serviceType, Driver* _driver) {
+    Bus::Bus(string _id, int _seats, vector<BusAmenities> _amenities, ServiceType _serviceType, Driver* _driver) {
         id = _id;
         seats = _seats;
         amenities = _amenities;
@@ -23,13 +23,13 @@ namespace tdc::domain::entities {
     };
 
     string Bus::getId() { return id; }
-    int8_t Bus::getSeats() { return seats; }
+    int Bus::getSeats() { return seats; }
     vector<BusAmenities> Bus::getAmenities() { return amenities; }
     ServiceType Bus::getServiceType() { return serviceType; }
     Driver* Bus::getDriver() { return driver; }
 
     void Bus::setId(string _id) { id = _id; };
-    void Bus::setSeats(int8_t _seats) { seats = _seats; };
+    void Bus::setSeats(int _seats) { seats = _seats; };
     void Bus::setAmenities(vector<BusAmenities> _amenities) { amenities = _amenities; };
     void Bus::setServiceType(ServiceType _serviceType) { serviceType = _serviceType; };
     void Bus::setDriver(Driver* _driver) { driver = _driver; };
